Updating classes after moving messages package out of the client package
