+++ zenvia-sdk-core/src/main/java/com/zenvia/api/sdk/client/Channel.java
@@ -11,10 +11,10 @@ import com.zenvia.api.sdk.client.exceptions.HttpIOException;
 import com.zenvia.api.sdk.client.exceptions.HttpSocketTimeoutException;
 import com.zenvia.api.sdk.client.exceptions.UnsuccessfulRequestException;
 import com.zenvia.api.sdk.client.exceptions.UnsupportedContentException;
-import com.zenvia.api.sdk.client.messages.MessageRequest;
 import com.zenvia.api.sdk.client.subscriptions.MessageStatusSubscription;
-import com.zenvia.api.sdk.client.messages.Message;
 import com.zenvia.api.sdk.contents.Content;
+import com.zenvia.api.sdk.messages.Message;
+import com.zenvia.api.sdk.messages.MessageRequest;
 
 
 /** Encapsulates the API calls under channel resource
