+++ zenvia-sdk-core/src/main/java/com/zenvia/api/sdk/messages/MessageDirection.java
@@ -1,4 +1,4 @@
-package com.zenvia.api.sdk.client.messages;
+package com.zenvia.api.sdk.messages;
 
 
 /** Indicates the origin of a message.
